api/requirements.txt
# -*- coding: utf-8 -*-
"""
AIVA API 要求檔案

定義 API 服務所需的 Python 套件依賴。
"""

# 核心 API 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 認證和安全
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# HTTP 客戶端
httpx==0.25.2
requests==2.31.0

# 數據驗證和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 系統監控（可選，如果需要系統資源監控）
psutil==5.9.6

# 異步支援
asyncio-mqtt==0.13.0

# 日誌記錄
structlog==23.2.0

# 開發和測試依賴
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-test==0.2.0